+++ sdk/core/Azure.Core.Expressions.DataFactory/src/DataFactoryElement.cs
@@ -22,7 +22,7 @@ namespace Azure.Core.Expressions.DataFactory
     {
         private readonly T? _literal;
         private readonly DataFactoryElementKind _kind;
-        internal DataFactorySecretDefinition? Secret { get; }
+        internal DataFactorySecret? Secret { get; }
         internal string? ExpressionString { get; }
 
         internal DataFactoryElement(T? literal)
@@ -56,7 +56,7 @@ namespace Azure.Core.Expressions.DataFactory
             ExpressionString = expressionString;
         }
 
-        internal DataFactoryElement(DataFactorySecretDefinition secret, DataFactoryElementKind kind)
+        internal DataFactoryElement(DataFactorySecret secret, DataFactoryElementKind kind)
         {
             _kind = kind;
             Secret = secret;
@@ -73,10 +73,10 @@ namespace Azure.Core.Expressions.DataFactory
             {
                 return ((DataFactorySecretString)Secret!).Value;
             }
-            if (_kind == DataFactoryElementKind.KeyVaultSecretReference)
+            if (_kind == DataFactoryElementKind.KeyVaultSecret)
             {
                 // TODO should this include the version and the Reference name?
-                return ((DataFactoryKeyVaultSecretReference)Secret!).SecretName.ToString();
+                return ((DataFactoryKeyVaultSecret)Secret!).SecretName.ToString();
             }
 
             return ExpressionString;
@@ -102,12 +102,12 @@ namespace Azure.Core.Expressions.DataFactory
         /// <summary>
         /// Creates a new instance of <see cref="DataFactoryElement{T}"/> using the KeyVaultSecretReference value.
         /// </summary>
-        /// <param name="secretReference"> The key vault secret reference value. </param>
+        /// <param name="secret"> The key vault secret reference value. </param>
 #pragma warning disable CA1000 // Do not declare static members on generic types
-        public static DataFactoryElement<string?> FromKeyVaultSecretReference(DataFactoryKeyVaultSecretReference secretReference)
+        public static DataFactoryElement<string?> FromKeyVaultSecret(DataFactoryKeyVaultSecret secret)
 #pragma warning restore CA1000 // Do not declare static members on generic types
         {
-            return new DataFactoryElement<string?>(secretReference, DataFactoryElementKind.KeyVaultSecretReference);
+            return new DataFactoryElement<string?>(secret, DataFactoryElementKind.KeyVaultSecret);
         }
 
         /// <summary>
@@ -126,7 +126,7 @@ namespace Azure.Core.Expressions.DataFactory
         /// </summary>
         /// <param name="secret"> The unmasked string value. </param>
 #pragma warning disable CA1000 // Do not declare static members on generic types
-        internal static DataFactoryElement<T?> FromSecretBase(DataFactorySecretDefinition secret)
+        internal static DataFactoryElement<T?> FromSecretBase(DataFactorySecret secret)
 #pragma warning restore CA1000 // Do not declare static members on generic types
         {
             return new DataFactoryElement<T?>(secret, new DataFactoryElementKind(secret.SecretBaseType!));